pyproject.toml
[tool.poetry]
name = "csb-project"
version = "0.1.0"
description = "cyber security base project 1"
authors = ["liisa <liisa-m@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.0"
flask-bootstrap = "^3.3.7.1"
flask-sqlalchemy = "^3.1.1"
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.0"
invoke = "^2.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
